rename: Nav로 변경(#45)
