+++ src/components/Sidebar/parts/index.ts
@@ -1,3 +1,3 @@
 export { default as DesktopAside } from './DesktopAside'
 export { default as MobileDrawer } from './MobileDrawer'
-export { default as SidebarNav } from './SidebarNav'
+export { default as Nav } from './Nav'
